Move TrajectoryGenerator and DriveOutput, fix comment
