+++ common/src/main/java/us/ilite/common/lib/trajectory/TrajectoryGenerator.java
@@ -1,8 +1,9 @@
-package us.ilite.common.lib;
+package us.ilite.common.lib.trajectory;
 
 import com.team254.lib.geometry.State;
 import com.team254.lib.trajectory.TrajectoryIterator;
 import com.team254.lib.util.Util;
+import us.ilite.common.lib.RobotProfile;
 import us.ilite.common.lib.control.DriveController;
 import us.ilite.common.lib.control.DriveMotionPlanner;
 import com.team254.lib.geometry.Pose2d;
